Move stray shared modules out of views/alerts
