+++ static/app/chartcuterie/metricAlert.tsx
@@ -1,15 +1,15 @@
 import type {Theme} from '@emotion/react';
 import type {LineSeriesOption, YAXisComponentOption} from 'echarts';
 
+import type {MetricChartData} from 'sentry/chartcuterie/metricChartOption';
+import {
+  getMetricAlertChartOption,
+  transformSessionResponseToSeries,
+} from 'sentry/chartcuterie/metricChartOption';
 import type {AreaChartSeries} from 'sentry/components/charts/areaChart';
 import {XAxis} from 'sentry/components/charts/components/xAxis';
 import {AreaSeries} from 'sentry/components/charts/series/areaSeries';
 import type {SessionApiResponse} from 'sentry/types/organization';
-import type {MetricChartData} from 'sentry/views/alerts/rules/metric/details/metricChartOption';
-import {
-  getMetricAlertChartOption,
-  transformSessionResponseToSeries,
-} from 'sentry/views/alerts/rules/metric/details/metricChartOption';
 
 import {DEFAULT_FONT_FAMILY, makeSlackChartDefaults, slackChartSize} from './slack';
 import type {RenderDescriptor} from './types';
