+++ static/app/chartcuterie/metricChartOption.tsx
