+++ static/app/views/automations/components/actions/sentryApp.tsx
@@ -6,8 +6,8 @@ import {useModal} from '@sentry/scraps/modal';
 import {IconSettings} from 'sentry/icons';
 import {t, tct} from 'sentry/locale';
 import type {ActionHandler} from 'sentry/types/workflowEngine/actions';
-import {SentryAppRuleModal} from 'sentry/views/alerts/rules/issue/sentryAppRuleModal';
 import {useActionNodeContext} from 'sentry/views/automations/components/actionNodes';
+import {SentryAppRuleModal} from 'sentry/views/automations/components/actions/sentryAppRuleModal';
 import type {SchemaFormConfig} from 'sentry/views/settings/organizationIntegrations/sentryAppExternalForm';
 
 export function SentryAppDetails({handler}: {handler: ActionHandler}) {
